+++ lib/page/feed/composeTweet/state/composeTweetState.dart
@@ -1,4 +1,11 @@
+import 'dart:convert';
+import 'package:flutter_twitter_clone/helper/utility.dart';
+import 'package:flutter_twitter_clone/model/user.dart';
+import 'package:http/http.dart' as http;
+import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_twitter_clone/helper/enum.dart';
+import 'package:flutter_twitter_clone/model/feedModel.dart';
 import 'package:flutter_twitter_clone/state/searchState.dart';
 
 class ComposeTweetState extends ChangeNotifier {
@@ -6,11 +13,12 @@ class ComposeTweetState extends ChangeNotifier {
   bool enableSubmitButton = false;
   bool hideUserList = false;
   String description = "";
+  String serverToken;
   final usernameRegex = r'(@\w*[a-zA-Z1-9]$)';
 
-  bool _isScrollingDown  = false;
-  bool get isScrollingDown  => _isScrollingDown;
-  set setIsScrolllingDown(bool value){
+  bool _isScrollingDown = false;
+  bool get isScrollingDown => _isScrollingDown;
+  set setIsScrolllingDown(bool value) {
     _isScrollingDown = value;
     notifyListeners();
   }
@@ -78,5 +86,88 @@ class ComposeTweetState extends ChangeNotifier {
     var name = description.substring(0, _matches.last.start);
     description = '$name $username';
     return description;
+  }
+
+  /// Fecth FCM server key from firebase Remote config
+  Future<Null> getFCMServerKey() async {
+    final RemoteConfig remoteConfig = await RemoteConfig.instance;
+    await remoteConfig.fetch(expiration: const Duration(hours: 5));
+    await remoteConfig.activateFetched();
+    var data = remoteConfig.getString('FcmServerKey');
+    if (data != null) {
+      serverToken = jsonDecode(data)["key"];
+    }
+  }
+
+  Future<void> sendNotification(FeedModel model, SearchState state) async {
+    final usernameRegex = r"(@\w*[a-zA-Z1-9])";
+    RegExp regExp = new RegExp(usernameRegex);
+    var status = regExp.hasMatch(description);
+    if (status) {
+      /// Fecth FCM server key from firebase Remote config
+      /// send notification to user once fcmToken is retrieved from firebase
+      getFCMServerKey().then((val) async {
+        /// Reset userlist
+        state.filterByUsername("");
+
+        /// Search all username from description
+        Iterable<Match> _matches = regExp.allMatches(description);
+        print("${_matches.length} name found in description");
+
+        /// Send notification to user one by one
+        await Future.forEach(_matches, (Match match) async {
+          var name = description.substring(match.start, match.end);
+          if (state.userlist.any((x) => x.userName == name)) {
+            /// Fetch user model from userlist
+            /// UserId, FCMtoken is needed to send notification
+            final user = state.userlist.firstWhere((x) => x.userName == name);
+            await sendNotificationToUser(model, user);
+          } else {
+            cprint("Name: $name ,", errorIn: "UserNot found");
+          }
+        });
+      });
+    }
+  }
+
+  /// Send notificatinn by using firebase notification rest api;
+  Future<void> sendNotificationToUser(FeedModel model, User user) async {
+    print("Send notification to: ${user.userName}");
+
+    /// Return from here if fcmToken is null
+    if (user.fcmToken == null) {
+      return;
+    }
+
+    /// Create notification payload
+    var body = jsonEncode(<String, dynamic>{
+      'notification': <String, dynamic>{
+        'body': model.description,
+        'title': "${model.user.displayName} metioned you in a tweet"
+      },
+      'priority': 'high',
+      'data': <String, dynamic>{
+        'click_action': 'FLUTTER_NOTIFICATION_CLICK',
+        'id': '1',
+        'status': 'done',
+        "type": NotificationType.Mention.toString(),
+        "senderId": model.user.userId,
+        "receiverId": user.userId,
+        "title": "title",
+        "body": "",
+        "tweetId": ""
+      },
+      'to': user.fcmToken
+    });
+
+    var response = await http.post(
+      'https://fcm.googleapis.com/fcm/send',
+      headers: <String, String>{
+        'Content-Type': 'application/json',
+        'Authorization': 'key=$serverToken',
+      },
+      body: body,
+    );
+    cprint(response.body.toString());
   }
 }